requirements.txt
fastapi==0.135.1
uvicorn==0.41.0
google-cloud-vision==3.12.1
google-auth==2.48.0
pillow==12.1.1
numpy==2.4.2
python-dotenv==1.2.2
pywebpush==2.3.0
py-vapid==1.9.4
python-multipart==0.0.22
supabase==2.28.0
